pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13.1"
dependencies = [
    "huggingface-hub>=0.29.1",
    "pydantic-ai>=0.0.24",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
